implement via abstract class
